Moved system.hpp and start_system.hpp into the systems folder.
