+++ core/include/bertini2/systems/start_system.hpp
