+++ core/include/bertini2/systems/system.hpp
