+++ core/src/system/Makemodule.am
@@ -11,5 +11,5 @@ system = $(system_header_files) $(system_source_files)
 
 
 rootinclude_HEADERS += \
-	include/bertini2/system.hpp \
-	include/bertini2/start_system.hpp
+	include/bertini2/systems/system.hpp \
+	include/bertini2/systems/start_system.hpp
